refactor createMapeoApi into new file
